Server/Misc: convert AccountTypes enum to scoped enum.

+++ src/server/scripts/Commands/cs_rbac.cpp
@@ -159,7 +159,7 @@ public:
 
         if (!rdata)
         {
-            data->rbac = new rbac::RBACData(accountId, accountName, realmID, AccountMgr::GetSecurity(accountId, realmID));
+            data->rbac = new rbac::RBACData(accountId, accountName, realmID, uint8(AccountMgr::GetSecurity(accountId, realmID)));
             data->rbac->LoadFromDB();
             data->needDelete = true;
         }
